+++ packages/assets/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next Release
 
-- Remove deprecated `fileType` API (#260).
+- Remove deprecated `fileType` API ([#260](https://github.com/andywer/webpack-blocks/issues/260))
 
 ## 1.0.0-rc