+++ packages/assets/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @webpack-blocks/assets - Changelog
 
-## Next Release
+## 2.0.0-alpha
 
 - Remove deprecated `fileType` API ([#260](https://github.com/andywer/webpack-blocks/issues/260))
 - Support for webpack 4 ([#261](https://github.com/andywer/webpack-blocks/pull/261))